+++ app/guid-file/controller.ts
@@ -4,6 +4,7 @@ import RouterService from '@ember/routing/router-service';
 import { inject as service } from '@ember/service';
 import { tracked } from '@glimmer/tracking';
 import { taskFor } from 'ember-concurrency-ts';
+import Intl from 'ember-intl/services/intl';
 import Media from 'ember-responsive';
 import Toast from 'ember-toastr/services/toast';
 
@@ -11,14 +12,17 @@ export default class GuidFile extends Controller {
     @service media!: Media;
     @service toast!: Toast;
     @service router!: RouterService;
+    @service intl!: Intl;
 
     @tracked revisionsOpened = false;
     @tracked tagsOpened = false;
+    @tracked metadataOpened = !this.isMobile;
+    @tracked resourceHelpOpen = false;
 
     @tracked viewedVersion?: number;
 
     get rightColumnClosed() {
-        return !(this.revisionsOpened || this.tagsOpened);
+        return !(this.revisionsOpened || this.tagsOpened || this.metadataOpened);
     }
 
     get isMobile() {
@@ -42,9 +46,13 @@ export default class GuidFile extends Controller {
         if (this.isMobile) {
             this.revisionsOpened = true;
             this.tagsOpened = false;
+            this.metadataOpened = false;
         } else {
             if (this.tagsOpened) {
                 this.tagsOpened = false;
+            }
+            if (this.metadataOpened) {
+                this.metadataOpened = false;
             }
             this.toggleProperty('revisionsOpened');
         }
@@ -55,11 +63,32 @@ export default class GuidFile extends Controller {
         if (this.isMobile) {
             this.tagsOpened = true;
             this.revisionsOpened = false;
+            this.metadataOpened = false;
         } else {
             if (this.revisionsOpened) {
                 this.revisionsOpened = false;
             }
+            if (this.metadataOpened) {
+                this.metadataOpened = false;
+            }
             this.toggleProperty('tagsOpened');
+        }
+    }
+
+    @action
+    toggleMetadata() {
+        if (this.isMobile) {
+            this.tagsOpened = false;
+            this.revisionsOpened = false;
+            this.metadataOpened = true;
+        } else {
+            if (this.revisionsOpened) {
+                this.revisionsOpened = false;
+            }
+            if (this.tagsOpened) {
+                this.tagsOpened = false;
+            }
+            this.toggleProperty('metadataOpened');
         }
     }
 
@@ -68,13 +97,16 @@ export default class GuidFile extends Controller {
         if (this.isMobile) {
             this.tagsOpened = false;
             this.revisionsOpened = false;
+            this.metadataOpened = false;
         } else {
             if (this.rightColumnClosed) {
-                this.revisionsOpened = true;
+                this.metadataOpened = true;
+                this.revisionsOpened = false;
                 this.tagsOpened = false;
             } else {
                 this.revisionsOpened = false;
                 this.tagsOpened = false;
+                this.metadataOpened = false;
             }
         }
     }