+++ app/models/custom-file-metadata-record.ts
@@ -1,0 +1,15 @@
+import { attr } from '@ember-data/model';
+import CustomMetadataModel from 'ember-osf-web/models/custom-metadata';
+
+
+export default class CustomFileMetadataRecordModel extends CustomMetadataModel {
+    @attr('fixstring') title?: string;
+    @attr('fixstring') description?: string;
+    @attr('fixstring') language?: string;
+}
+
+declare module 'ember-data/types/registries/model' {
+    export default interface ModelRegistry {
+        'custom-file-metadata-record': CustomFileMetadataRecordModel;
+    } // eslint-disable-line semi
+}