+++ app/router.ts
@@ -62,6 +62,7 @@ Router.map(function() {
         this.route('files', function() {
             this.route('provider', { path: '/:providerId' });
         });
+        this.route('metadata');
         this.route('registrations');
         this.route('drafts', { path: '/drafts/:draftId' }, function() {
             this.route('register');