+++ lib/registries/addon/overview/controller.ts
@@ -24,7 +24,7 @@ const {
     },
     OSF: {
         cookies: {
-            outputFeaturePopover,
+            newFeaturePopover,
         },
     },
 } = config;
@@ -41,7 +41,7 @@ export default class Overview extends Controller {
 
     queryParams = ['mode', 'revisionId'];
     supportEmail = supportEmail;
-    outputFeaturePopoverCookie = outputFeaturePopover;
+    newFeaturePopoverCookie = newFeaturePopover;
 
     @tracked mode = '';
     @tracked revisionId = '';
@@ -59,7 +59,8 @@ export default class Overview extends Controller {
     get showMetadata() {
         if (
             this.router.currentRouteName.includes('registries.overview.files') ||
-            this.router.currentRouteName.includes('registries.overview.children')
+            this.router.currentRouteName.includes('registries.overview.children') ||
+            this.router.currentRouteName.includes('registries.overview.metadata')
         ) {
             return false;
         }