+++ mirage/factories/node.ts
@@ -155,7 +155,7 @@ export default Factory.extend<MirageNode & NodeTraits>({
 
     withFiles: trait<MirageNode>({
         afterCreate(node, server) {
-            const count = faker.random.number({ min: 1, max: 5 });
+            const count = 3;
             const providerId = node.id + ':osfstorage';
             const osfstorage = server.create('file-provider', { id: providerId, target: node });
             const files = server.createList('file', count, { target: node });