+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [23.03.0] - 2023-02-06
+### Changed
+- Collections Moderation bug and improvements
+
+## [23.02.1] - 2023-01-31
+### Changed
+- Hotfix the bug where metadata does not display for files on providers that have neither tags nor revisions
+
+## [23.02.0] - 2023-01-30
+### Added
+- GUID metadata
+
 ## [23.01.0] - 2023-01-25
 ### Changed
 - Collections Moderation bug and improvements
@@ -1889,6 +1901,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Quick Files
 
+[23.03.0]: https://github.com/CenterForOpenScience/ember-osf-web/releases/tag/23.03.0
+[23.02.1]: https://github.com/CenterForOpenScience/ember-osf-web/releases/tag/23.02.1
+[23.02.0]: https://github.com/CenterForOpenScience/ember-osf-web/releases/tag/23.02.0
 [23.01.0]: https://github.com/CenterForOpenScience/ember-osf-web/releases/tag/23.01.0
 [22.11.0]: https://github.com/CenterForOpenScience/ember-osf-web/releases/tag/22.11.0
 [22.10.0]: https://github.com/CenterForOpenScience/ember-osf-web/releases/tag/22.10.0